refactor: centralize TypeScript type definitions for better organization
- Create dedicated src/types/ directory with organized type definitions
- Add src/types/environment.ts for environment configuration types
- Add src/types/parameter-store.ts for Parameter Store related interfaces
- Add src/types/lambda.ts for Lambda function result types
- Add src/types/index.ts barrel file for convenient importing
- Update all implementation files to use centralized types
- Remove scattered interface definitions from implementation files
- Maintain backward compatibility and test coverage

Benefits:
- Better TypeScript code organization and maintainability
- Centralized type definitions prevent duplication
- Easy importing with: import { TypeName } from '../types'
- Improved IDE support and type checking

Resolves SuggestedChanges725 issue #4 (MEDIUM priority)
