+++ src/config/environment.ts
@@ -1,8 +1,4 @@
-export interface EnvironmentConfig {
-  environment: 'development' | 'production' | 'test';
-  awsRegion: string;
-  awsAccount: string;
-}
+import { EnvironmentConfig } from '../types/environment';
 
 const getEnvironmentConfig = (): EnvironmentConfig => {
   const environment = (process.env.NODE_ENV || 'development') as 'development' | 'production' | 'test';
