+++ src/types/environment.ts
@@ -1,0 +1,9 @@
+/**
+ * Environment configuration types
+ */
+
+export interface EnvironmentConfig {
+  environment: 'development' | 'production' | 'test';
+  awsRegion: string;
+  awsAccount: string;
+}
