+++ src/types/index.ts
@@ -1,0 +1,15 @@
+/**
+ * Central type definitions for the Report Builder application
+ * 
+ * This barrel file exports all types for convenient importing:
+ * import { EnvironmentConfig, EmailProcessorResult } from '../types';
+ */
+
+// Environment types
+export * from './environment';
+
+// Parameter Store types  
+export * from './parameter-store';
+
+// Lambda function types
+export * from './lambda';
